NH_voice_agent/3_voice_app/requirements.txt
# Databricks SDK
databricks-sdk>=0.20.0

# Voice Processing (excluding PyAudio and Whisper for cloud deployment)
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
pydub>=0.25.1

# Web Framework
streamlit>=1.28.0
audio-recorder-streamlit>=0.0.8

# Data Processing
pandas>=2.0.0

# LangChain for Agent
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
